Updates app with reset, winner display, and auth
Updated the app by setting up a reset button that can reset the game at
any point. Also added a display above the board that displays the winner
at the end of the game, also if there is a draw. Lastly, I added the
stock auth code from api-token-auth that contains sign up, sign in,
and change password functionality.